requirements.txt
# Minimal runtime requirements for CLI crawler
# Optional: load .env without breaking when missing
python-dotenv>=1.0.0
